pyproject.toml
[project]
name = "chexformer"
version = "1.0.0"
description = "Final Project for Electronics and Computer Engineering @ UFRJ - Lucas Cruz."
authors = [
    {name = "Lucas Cruz", email = "lucasdearaujocruz@gmail.com"},
    {name = "Lucas Cruz"},
]

requires-python = "==3.10"
readme = "README.md"
license = {text = "MIT License"}
dependencies = [
    "accelerate>=0.31.0",
    "google-cloud-storage>=2.17.0",
    "huggingface-hub>=0.23.4",
    "ipykernel>=6.29.4",
    "ipython>=8.18.1",
    "jupyterlab>=4.2.2",
    "matplotlib>=3.9.0",
    "pandas>=1.5.2",
    "numpy>=1.23.5",
    "pillow>=10.3.0",
    "transformers>=4.41.2",
    "torchmetrics>=1.4.0.post0",
    "torchvision>=0.18.1",
    "webdataset>=0.2.86",
    "tqdm>=4.66.4",
    "setuptools>=70.1.0",
    "pip>=24.0",
    "wheel>=0.43.0",
    "dataclasses-json>=0.6.7",
    "gcsfs>=2024.6.0",
    "mlflow>=2.14.1",
    "peft>=0.11.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = ["-e file:///${PROJECT_ROOT}/#egg=chexformer"]
